- create following files in root directory 	- first_time_setup.sh 	- quick_start.sh 	- pause_dockers.sh 	- remove_dockers.sh - Renamed all first_time_setup.sh in internal folders and their references - Renamed all quick_start.sh in internal folders and their references

+++ packages/docker-eosio-nodeos/start_eosio_docker.sh
@@ -4,6 +4,11 @@ set -o errexit
 # change to script's directory
 cd "$(dirname "$0")"
 
+# docker did not stop properly
+if [ "$(docker ps -q -f status=exited -f name=eosio_gui_nodeos_container)" ]; then
+  docker rm eosio_gui_nodeos_container
+fi
+
 if [ -e "data/initialized" ]
 then
     script="./scripts/continue_blockchain.sh"
@@ -11,18 +16,22 @@ else
     script="./scripts/init_blockchain.sh"
 fi
 
-# --link is to get access to other container
-echo "=== run docker container from the eosio-gui-nodeos:eos1.6.3 image ==="
-docker run --rm --name eosio_gui_nodeos_container -d \
--p 8888:8888 -p 9876:9876 \
---link eosio-mongodb \
---mount type=bind,src="$(pwd)"/contracts,dst=/opt/eosio/bin/contracts \
---mount type=bind,src="$(pwd)"/scripts,dst=/opt/eosio/bin/scripts \
---mount type=bind,src="$(pwd)"/data,dst=/mnt/dev/data \
--w "/opt/eosio/bin/" eosio-gui-nodeos:eos1.6.3 /bin/bash -c "$script"
+if [ ! "$(docker ps -q -f name=eosio_gui_nodeos_container)" ]; then
+    # --link is to get access to other container
+    echo "=== run docker container from the eosio-gui-nodeos:eos1.6.3 image ==="
+    docker run --rm --name eosio_gui_nodeos_container -d \
+    -p 8888:8888 -p 9876:9876 \
+    --link eosio-mongodb \
+    --mount type=bind,src="$(pwd)"/contracts,dst=/opt/eosio/bin/contracts \
+    --mount type=bind,src="$(pwd)"/scripts,dst=/opt/eosio/bin/scripts \
+    --mount type=bind,src="$(pwd)"/data,dst=/mnt/dev/data \
+    -w "/opt/eosio/bin/" eosio-gui-nodeos:eos1.6.3 /bin/bash -c "$script"
 
-if [ "$1" != "--nolog" ]
-then
-    echo "=== follow eosio_gui_nodeos_container logs ==="
-    docker logs eosio_gui_nodeos_container --follow
+    if [ "$1" != "--nolog" ]
+    then
+        echo "=== follow eosio_gui_nodeos_container logs ==="
+        docker logs eosio_gui_nodeos_container --follow
+    fi
+else
+    echo "docker already running"
 fi
